11 проект, 1 итерация.

+++ src/utils/Api.js
@@ -64,18 +64,18 @@ class Api {
     }).then(this._checkResponse);
   }
 
-  putLike(cardId) {
-    return fetch(`${this._address}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._checkResponse);
-  }
-
-  deleteLike(cardId) {
-    return fetch(`${this._address}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._checkResponse);
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._address}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    } else {
+      return fetch(`${this._address}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    }
   }
 
   getUsers() {
@@ -86,12 +86,10 @@ class Api {
   }
 }
 
-const api = new Api({
+export const api = new Api({
   baseUrl: "https://mesto.nomoreparties.co/v1/cohort-64",
   headers: {
     authorization: "f64efd3e-3c8d-4834-a748-041ca1b1af96",
     "Content-Type": "application/json",
   },
 });
-
-export default api;
